Use `String#append_bytes` if available
This is just a proof of concept / demo, there are some unknown
about how codegen is supposed to know if it's OK to use newly introduced
methods.

I also hacked the benchmark to load two versions of protoboeuf so
I can compare them together, that's proable not how we want it
but it gives a much clearer picture of the speedup.

```
=== encode ===
ruby 3.4.0dev (2024-08-02T20:36:59Z string-append-bytes 1eeb922be5) +YJIT [arm64-darwin23]
Warming up --------------------------------------
      protoboeuf/jit     4.000 i/100ms
        upstream/jit    13.000 i/100ms
     pboeuf-edge/jit     4.000 i/100ms
Calculating -------------------------------------
      protoboeuf/jit     40.731 (± 0.0%) i/s -    204.000 in   5.009091s
        upstream/jit    134.985 (± 1.5%) i/s -    676.000 in   5.009576s
     pboeuf-edge/jit     45.224 (± 2.2%) i/s -    228.000 in   5.043503s

Comparison:
      protoboeuf/jit:       40.7 i/s
        upstream/jit:      135.0 i/s - 3.31x  faster
     pboeuf-edge/jit:       45.2 i/s - 1.11x  faster
```